# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TeliTask Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: telegramLogin
      tags: [auth]
      summary: Login or register with Telegram initData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid initData
...
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [completions]
      summary: Submit task completion
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionInput"
      responses:
        "201":
          description: Completion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Completion"
        "400":
          description: Already completed or time expired
...
  /completions:
    get:
      operationId: listAllCompletions
      tags: [completions]
      summary: List all completions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Completions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionListResult"
  /completions/{id}/approve:
    patch:
      operationId: approveCompletion
      tags: [completions]
      summary: Approve a task completion (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Completion"
  /completions/{id}/reject:
    patch:
      operationId: rejectCompletion
      tags: [completions]
      summary: Reject a task completion (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Completion"
...
  /notifications/broadcast:
    post:
      operationId: broadcastNotification
      tags: [notifications]
      summary: Send broadcast notification to all users (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResult"
  /notifications/personal:
    post:
      operationId: sendPersonalNotification
      tags: [notifications]
      summary: Send personal notification to a specific user by Telegram ID (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalNotifInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResult"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    TelegramLoginInput:
      type: object
      required: [initData]
      properties:
        initData:
          type: string
          description: Telegram WebApp initData string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, telegramId, username, firstName, balance, isBanned, isAdmin, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        balance:
          type: number
        totalEarned:
          type: number
        totalWithdrawn:
          type: number
        completedTasksCount:
          type: integer
        isBanned:
          type: boolean
        isAdmin:
          type: boolean
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Task:
      type: object
      required: [id, title, description, reward, timeLimitMinutes, isActive, requiresScreenshot, requiresLinkCopy, completionCount, dailyLimit, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: number
        timeLimitMinutes:
          type: integer
          description: Time limit in minutes to complete the task
        isActive:
          type: boolean
        requiresScreenshot:
          type: boolean
        requiresLinkCopy:
          type: boolean
        copyLink:
          type: ["string", "null"]
          description: Link to copy for the task
        dailyLimit:
          type: integer
          description: Max completions per user per day (0 = once total)
        completionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        userCompleted:
          type: ["boolean", "null"]
          description: Whether current user already completed this task
...
    Completion:
      type: object
      required: [id, userId, taskId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        taskId:
          type: integer
        task:
          $ref: "#/components/schemas/Task"
        user:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [pending, approved, rejected]
        screenshotUrl:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        reward:
          type: number
        createdAt:
          type: string
          format: date-time
    CompletionInput:
      type: object
      properties:
        screenshotUrl:
          type: string
        note:
          type: string
    CompletionListResult:
      type: object
      required: [completions, total, page]
      properties:
        completions:
          type: array
          items:
            $ref: "#/components/schemas/Completion"
        total:
          type: integer
        page:
          type: integer
    RejectionInput:
      type: object
      properties:
        reason:
          type: string
...
    Notification:
      type: object
      required: [id, userId, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        type:
          type: string
          enum: [broadcast, personal, system]
        createdAt:
          type: string
          format: date-time
    BroadcastInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string
    PersonalNotifInput:
      type: object
      required: [telegramId, title, message]
      properties:
        telegramId:
          type: string
        title:
          type: string
        message:
          type: string
    NotificationResult:
      type: object
      required: [sent]
      properties:
        sent:
          type: integer
...